+++ NEWS.md
@@ -1,3 +1,15 @@
+# APfun 0.1.2
+
+Added following functions to facilitate reading and writing of Shapefiles
+
+APSHPread - Read a Shapefile
+APSHPsave - Save a Shapefile
+APKML2SHP - Convert a KML to a Shapefile
+APSHPfiles - List all files associated with a Shapefile (ex.: yourfile.shp, yourfile.xml, yourfile.prj, etc.)
+APSHPdel - Delete Shapefile and all associated files
+
+In addition, the 'APpolyMerge' function was added, which facilitates merging multiple polygonal Spatial objects.
+
 # APfun 0.1.1
 
 Added following functions: